% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{clr_coordinates}
\alias{clr_coordinates}
\title{clr coordinates of compositional data}
\usage{
clr_coordinates(X)
}
\arguments{
\item{X}{compositional sample}
}
\description{
clr coordinates of compositional data
}

