% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_function.R
\name{clr_mixnorm}
\alias{clr_mixnorm}
\title{CLR evaluated on gaussian mixture model posterioris of X}
\usage{
clr_mixnorm(X, Pi, Mu, Sigma)
}
\arguments{
\item{X}{dataframe where density function is evaluated}

\item{Pi}{a vector indicating the mixing proportions}

\item{Mu}{a two dimensional array where second component indicates the mean 
of each gaussian component}

\item{Sigma}{a three dimensional array where third component indicates the variance 
of each gaussian component}
}
\description{
CLR evaluated on gaussian mixture model posterioris of X
}

