% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combining_partitions.R
\name{cluster_partition}
\alias{cluster_partition}
\title{Create a cluster from a partition}
\usage{
cluster_partition(tau, partition)
}
\arguments{
\item{tau}{matrix of posterioris}

\item{partition}{list of vectors containing the partition}
}
\description{
Given a matrix of tau and a partition decide in which part is classified each observation
}

