% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mixture_function.R
\name{dmixnorm_solution}
\alias{dmixnorm_solution}
\title{Density function of specified gaussian mixture model.}
\usage{
dmixnorm_solution(x, solution, ...)
}
\arguments{
\item{x}{vector/matrix where density function is evaluated}

\item{solution}{solution comming from packages \code{Mclust} or \code{rmixmod}}

\item{...}{arguments passed to function \code{\link{dmixnorm}}}
}
\description{
The parameters are defined from the parameters obtained using other
packages (' \code{Mclust},
\code{rmixmod})
}
\examples{
require(mclust)
mod1 = Mclust(iris[,1:4])
rmixnorm_solution(10, mod1)
}

