% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_hierarchical_partition_mult_fast}
\alias{get_hierarchical_partition_mult_fast}
\title{Build a hierchical partition from posterior probabilities}
\usage{
get_hierarchical_partition_mult_fast(post, omega = "prop", lambda = "coda")
}
\arguments{
\item{omega}{function with two parameters (\code{v_tau}, \code{a}). Parameter
\code{v_tau} is a vector of probabilities, parameter \code{a} is the a selected class.
\code{omega}(\code{v_tau}, \code{a}) gives the representativeness of element with
probabities \code{v_tau} to class \code{a}}

\item{lambda}{function with three parameters (\code{v_tau}, \code{a}, \code{b}).
Parameter \code{v_tau} is a vector of probabilities, parameters \code{a} and \code{b}
are classes to be combined.}

\item{tau}{dataframe of probabilities/weights (\code{tau} must be strictly positive)}
}
\description{
This function applies the methodology described in [citar article]
to build a hierarchy of classes using the weights or probabilities
that an element belongs to each class
}

