% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combining_partitions.R
\name{get_hierarchical_partition_mult_func}
\alias{get_hierarchical_partition_mult_func}
\title{Build a hierchical partition from posterior probabilities}
\usage{
get_hierarchical_partition_mult_func(tau, varphi, theta, func = function(tau,
  partition) do.call("cbind", lapply(partition, function(part) {     if
  (is.vector(tau[, part]))          return(tau[, part])     apply(tau[, part],
  1, prod)^(1/length(part)) })))
}
\arguments{
\item{tau}{dataframe of probabilities/weights (\code{tau} must be strictly positive)}

\item{func}{defines the way components are merged}

\item{omega}{function with two parameters (\code{v_tau}, \code{a}). Parameter
\code{v_tau} is a vector of probabilities, parameter \code{a} is the a selected class.
\code{omega}(\code{v_tau}, \code{a}) gives the representativeness of element with
probabities \code{v_tau} to class \code{a}}

\item{lambda}{function with three parameters (\code{v_tau}, \code{a}, \code{b}).
Parameter \code{v_tau} is a vector of probabilities, parameters \code{a} and \code{b}
are classes to be combined.}
}
\description{
This function applies the methodology described in [citar article]
to build a hierarchy of classes using the weights or probabilities
that an element belongs to each class
}

