% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combining_partitions.R
\name{get_random_hierarchical_partition}
\alias{get_random_hierarchical_partition}
\title{Build a hierchical partition randomly from given K}
\usage{
get_random_hierarchical_partition(K)
}
\arguments{
\item{K}{number of initial groups}
}
\description{
This function return a hierachical partition contructed randonmly.
}

