% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining_partitions.R
\name{merge_components}
\alias{merge_components}
\title{Build a hierchical partition randomly from given K}
\usage{
merge_components(post, a, b)
}
\arguments{
\item{post}{posterior probability matrix}

\item{a}{first component to merge}

\item{b}{second component to merge}
}
\value{
a matrix of posterior probabilities where components a and b are merged
}
\description{
This function return a hierachical partition contructed randonmly.
}

