% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining_partitions.R
\name{merge_step}
\alias{merge_step}
\title{Merging components step}
\usage{
merge_step(post, omega, lambda, f_omega = NULL, f_lambda = NULL)
}
\arguments{
\item{post}{Matrix with the posterior probabilities}

\item{omega}{String giving the function name used to build the hierarchy. Available 
functions are: entr, prop, dich}

\item{lambda}{String giving the function name used to build the hierarchy. Available 
functions are: entr, demp, demp.mod, coda, coda.norm, prop}

\item{f_omega}{function with two parameters (\code{v_tau}, \code{a}). Parameter 
\code{v_tau} is a vector of probabilities, parameter \code{a} is the a selected class.
\code{omega}(\code{v_tau}, \code{a}) gives the representativeness of element with
probabities \code{v_tau} to class \code{a}}

\item{f_lambda}{function with three parameters (\code{v_tau}, \code{a}, \code{b}).
Parameter \code{v_tau} is a vector of probabilities, parameters \code{a} and \code{b}
are classes to be combined.}
}
\value{
partition returns a matrix with all values for all possible mergings using functions `omega` and `lambda`
}
\description{
Merging components step
}

