% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mergeStep}
\alias{mergeStep}
\title{Merging components step}
\usage{
mergeStep(post, omega = "prop", lambda = "coda")
}
\arguments{
\item{post}{Matrix with the posterior probabilities}

\item{omega}{omega function name}

\item{lambda}{lambda function name}
}
\value{
partition using prop and codaNorm
}
\description{
Merging components step
}

