% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mergeStep_const_entropy}
\alias{mergeStep_const_entropy}
\title{Merging components step}
\usage{
mergeStep_const_entropy(post)
}
\arguments{
\item{post}{Matrix with the posterior probabilities}
}
\value{
partition using const and entropy
}
\description{
Merging components step
}

