% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mergeStep_prop_codaNorm}
\alias{mergeStep_prop_codaNorm}
\title{Merging components step}
\usage{
mergeStep_prop_codaNorm(post)
}
\arguments{
\item{post}{Matrix with the posterior probabilities}
}
\description{
Merging components step
}

