% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmixnorm}
\alias{rmixnorm}
\title{Random sample generated from an specified gaussian mixture model.}
\usage{
rmixnorm(n, Pi, Mu, S, labels = F)
}
\arguments{
\item{n}{sample size}

\item{Pi}{a vector indicating the mixing proportions}

\item{Mu}{a two dimensional array where second component indicates the mean
of each gaussian component}

\item{S}{a three dimensional array where third component indicates the variance
of each gaussian component}

\item{labels}{a boolean indicating whether or not a label shoud be returned indicating
the component from where observation has been generated}
}
\description{
Random sample generated from an specified gaussian mixture model.
}

