% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmixnorm_solution}
\alias{rmixnorm_solution}
\title{Random sample generated from an specified gaussian mixture model.}
\usage{
rmixnorm_solution(n, solution, ...)
}
\arguments{
\item{n}{sample size}

\item{solution}{solution comming from packages \code{Mclust} or \code{rmixmod}}
}
\description{
The parameters are defined from the parameters obtained using other
packages (' \code{Mclust},
\code{rmixmod})
}
\examples{
require(mclust)
mod1 = Mclust(iris[,1:4])
rmixnorm_solution(10, mod1)
}

