% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dmixnorm}
\alias{dmixnorm}
\title{Density function of specified gaussian mixture model.}
\usage{
dmixnorm(x, Pi, Mu, S, part = 1:length(Pi))
}
\arguments{
\item{x}{vector/matrix where density function is evaluated}

\item{Pi}{a vector indicating the mixing proportions}

\item{Mu}{a two dimensional array where second component indicates the mean
of each gaussian component}

\item{S}{a three dimensional array where third component indicates the variance
of each gaussian component}

\item{part}{subcomposition where x shoud be evaluated. Take into an account that
if x has dimensions K, K components must be selected by \code{part}}
}
\description{
Density function of specified gaussian mixture model.
}

