% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coda.R
\name{ilr_basis}
\alias{ilr_basis}
\title{Basis from the simplex space with $D$ components}
\usage{
ilr_basis(D)
}
\arguments{
\item{D}{numbre of components}
}
\description{
Basis from the simplex space with $D$ components
}

