% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coda.R
\name{ilr_coordinates}
\alias{ilr_coordinates}
\title{Coordinates for an orthonormal basis

Coordinates respect basis \code{\link{ilr_basis}}}
\usage{
ilr_coordinates(X)
}
\arguments{
\item{X}{compositional sample}
}
\description{
Coordinates for an orthonormal basis

Coordinates respect basis \code{\link{ilr_basis}}
}

